--- web/src/app/(app)/events/styles.css
@@ -1,27 +1,0 @@
-div {
-    background-color: #fff2d1;
-    padding: 20px 40px;
-    margin: 0;
-  }
-  
-  table {
-    text-align: center;
-    width: 100%;
-    border-spacing: 0;
-    table-layout: fixed;
-    border: solid 1px #000;
-  }
-  
-  thead {
-    text-align: center;
-  }
-  
-  th,td {
-    width: 80%;
-    table-layout: fixed;
-    border:1.5px #7e7e7e solid;
-  }
-
-  
-  
-
